+++ crates/spam-filter/src/analysis/from.rs
@@ -35,7 +35,7 @@ impl SpamFilterAnalyzeFrom for Server {
                         .input
                         .message
                         .raw_message()
-                        .get(header.offset_start..header.offset_end)
+                        .get(header.offset_start as usize..header.offset_end as usize)
                         .unwrap_or_default();
                 }
                 HeaderName::Other(name) => {
