+++ crates/spam-filter/src/analysis/recipient.rs
@@ -35,7 +35,7 @@ impl SpamFilterAnalyzeRecipient for Server {
                         .input
                         .message
                         .raw_message()
-                        .get(header.offset_start..header.offset_end)
+                        .get(header.offset_start as usize..header.offset_end as usize)
                         .unwrap_or_default();
                     match header.name {
                         HeaderName::To => to_raw = raw,
